Add bandos godsword special attack
