+++ data/definitions/animation-definition-extras.yml
@@ -1307,3 +1307,5 @@ backstab:
   id: 4198
 sunder:
   id: 5870
+warstrike:
+  id: 11991
