+++ data/definitions/graphic-definition-extras.yml
@@ -804,3 +804,5 @@ backstab:
   id: 704
 sunder:
   id: 1027
+warstrike:
+  id: 2114
