+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/SpecialAttackMelee.kt
@@ -1,9 +1,12 @@
 package world.gregs.voidps.world.interact.entity.player.combat.melee
 
+import world.gregs.voidps.engine.entity.character.Character
 import world.gregs.voidps.engine.entity.character.player.Player
+import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
+import world.gregs.voidps.world.interact.entity.combat.CombatHit
 import world.gregs.voidps.world.interact.entity.combat.HitDamageModifier
 import world.gregs.voidps.world.interact.entity.combat.HitRatingModifier
 import kotlin.math.floor
@@ -16,5 +19,24 @@ fun specialDamageMultiplier(multiplier: Double, check: (Item) -> Boolean) {
 fun specialAccuracyMultiplier(multiplier: Double, check: (Item) -> Boolean) {
     on<HitRatingModifier>({ offense && type == "melee" && special && weapon != null && check(weapon) }, Priority.HIGH) { _: Player ->
         rating = floor(rating * multiplier)
+    }
+}
+
+fun specialDrainByDamage(check: (Item) -> Boolean, vararg skills: Skill) {
+    on<CombatHit>({ weapon != null && special && check(weapon) }) { character: Character ->
+        var drain = damage / 10
+        if (drain > 0) {
+            for (skill in skills) {
+                val current = character.levels.get(skill)
+                if (current <= 1) {
+                    continue
+                }
+                character.levels.drain(skill, drain)
+                drain -= current
+                if (drain <= 0) {
+                    break
+                }
+            }
+        }
     }
 }
