+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/special/BoneDagger.kts
@@ -1,6 +1,5 @@
 package world.gregs.voidps.world.interact.entity.player.combat.melee.special
 
-import world.gregs.voidps.engine.entity.character.Character
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.entity.character.update.visual.setAnimation
@@ -8,7 +7,11 @@ import world.gregs.voidps.engine.entity.character.update.visual.setGraphic
 import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.interact.entity.combat.*
+import world.gregs.voidps.world.interact.entity.combat.CombatSwing
+import world.gregs.voidps.world.interact.entity.combat.HitChanceModifier
+import world.gregs.voidps.world.interact.entity.combat.hit
+import world.gregs.voidps.world.interact.entity.combat.weapon
+import world.gregs.voidps.world.interact.entity.player.combat.melee.specialDrainByDamage
 import world.gregs.voidps.world.interact.entity.player.combat.range.special.drainSpecialEnergy
 import world.gregs.voidps.world.interact.entity.player.combat.range.special.specialAttack
 
@@ -30,9 +33,4 @@ on<CombatSwing>({ !swung() && it.specialAttack && isBoneDagger(it.weapon) }) { p
     delay = 4
 }
 
-on<CombatHit>({ isBoneDagger(weapon) && special }) { character: Character ->
-    val drain = damage / 10
-    if (drain > 0) {
-        character.levels.drain(Skill.Defence, drain)
-    }
-}
+specialDrainByDamage(::isBoneDagger, Skill.Defence)
